重构Model
