+++ src/model/concern/Attribute.php
@@ -22,7 +22,11 @@ use think\db\Raw;
 use think\helper\Str;
 use think\model\contract\EnumTransform;
 use think\model\contract\FieldTypeTransform;
-use think\model\Relation;
+use think\model\contract\Typeable;
+use think\model\contract\Modelable;
+use think\model\type\Date;
+use think\model\type\DateTime;
+use think\model\type\Json;
 
 /**
  * 模型数据处理.
@@ -30,281 +34,91 @@ use think\model\Relation;
 trait Attribute
 {
     /**
-     * 数据表主键 复合主键使用数组定义.
+     * 初始化模型数据.
      *
-     * @var string|array
-     */
-    protected $pk = 'id';
-
-    /**
-     * 数据表主键自增.
+     * @param array|object $data 实体模型数据
+     * @param bool  $fromSave
      *
-     * @var bool|null|string
+     * @return void
      */
-    protected $autoInc;
+    private function initializeData(array | object $data, bool $fromSave = false)
+    {
+        // 分析数据
+        $data    = $this->parseData($data);
+        $schema  = $this->getFields();
+        $fields  = array_keys($schema);
+        $mapping = $this->getOption('mapping');
 
-    /**
-     * 数据表字段信息 留空则自动获取.
-     *
-     * @var array
-     */
-    protected $schema = [];
+        // 实体模型赋值
+        foreach ($data as $name => $val) {
+            if (in_array($name, $this->getOption('disuse'))) {
+                // 废弃字段
+                continue;
+            }
 
-    /**
-     * 当前允许写入的字段.
-     *
-     * @var array
-     */
-    protected $field = [];
+            if (!empty($mapping)) {
+                $name = array_search($name, $mapping) ?: $name;
+            }
 
-    /**
-     * 字段自动类型转换.
-     *
-     * @var array
-     */
-    protected $type = [];
+            if (str_contains($name, '__')) {
+                // 组装关联JOIN查询数据
+                [$relation, $attr] = explode('__', $name, 2);
 
-    /**
-     * 数据表废弃字段.
-     *
-     * @var array
-     */
-    protected $disuse = [];
+                $relations[$relation][$attr] = $val;
+                continue;
+            }
 
-    /**
-     * 数据表只读字段.
-     *
-     * @var array
-     */
-    protected $readonly = [];
+            $trueName = $this->getRealFieldName($name);
+            if ($this->isView() || $this->isVirtual() || in_array($trueName, $fields)) {
+                // 读取数据后进行类型转换
+                $value = $this->readTransform($val, $schema[$trueName] ?? 'string');
+                // 数据赋值
+                $this->setData($trueName, $value);
+                // 记录原始数据
+                $origin[$trueName] = $value;
+            }
+        }
 
-    /**
-     * 当前模型数据.
-     *
-     * @var array
-     */
-    private $data = [];
+        if (!empty($relations)) {
+            // 设置关联数据
+            $this->parseRelationData($relations);
+        }
 
+        if (!empty($origin) && !$fromSave) {
+            $this->trigger('AfterRead');
+            $this->setOption('origin', $origin);
+            $this->setOption('get', []);
+        }
+    }
+        
     /**
-     * 原始数据.
-     *
-     * @var array
-     */
-    private $origin = [];
-
-    /**
-     * JSON数据表字段.
-     *
-     * @var array
-     */
-    protected $json = [];
-
-    /**
-     * JSON数据表字段类型.
-     *
-     * @var array
-     */
-    protected $jsonType = [];
-
-    /**
-     * JSON数据取出是否需要转换为数组.
-     *
-     * @var bool
-     */
-    protected $jsonAssoc = false;
-
-    /**
-     * Enum数据取出自动转换为name.
-     *
-     * @var bool|string
-     */
-    protected $enumReadName = false;
-
-    /**
-     * 严格检查Enum数据类型.
-     *
-     * @var bool
-     */
-    protected $enumStrict = false;
-
-    /**
-     * 是否严格字段大小写.
-     *
-     * @var bool
-     */
-    protected $strict = true;
-
-    /**
-     * 获取器数据.
-     *
-     * @var array
-     */
-    private $get = [];
-
-    /**
-     * 动态获取器.
-     *
-     * @var array
-     */
-    private $withAttr = [];
-
-    /**
-     * 自动写入字段.
-     *
-     * @var array
-     */
-    protected $insert = [];
-
-    /**
-     * 主键值.
-     *
-     * @var int|string
-     */
-    protected $key;
-
-    /**
-     * 获取模型对象的主键.
+     * 获取主键名.
      *
      * @return string|array
      */
     public function getPk()
     {
-        return $this->pk;
+        return $this->getOption('pk');
     }
 
     /**
-     * 设置模型的字段信息.
-     *
-     * @param array $schema
-     * @return $this
-     */
-    public function schema(array $schema)
-    {
-        $this->schema = $schema;
-        return $this;
-    }
-
-    public function isJsonAssoc(): bool
-    {
-        return $this->jsonAssoc;
-    }
-
-    public function getType()
-    {
-        return $this->type;
-    }
-
-    /**
-     * 判断一个字段名是否为主键字段.
-     *
-     * @param string $key 名称
-     *
-     * @return bool
-     */
-    protected function isPk(string $key): bool
-    {
-        $pk = $this->getPk();
-
-        if (is_string($pk) && $pk == $key) {
-            return true;
-        } elseif (is_array($pk) && in_array($key, $pk)) {
-            return true;
-        }
-
-        return false;
-    }
-
-    public function setKey($value)
-    {
-        $pk = $this->getPk();
-
-        $this->key    = $value;
-        $this->exists = true;
-
-        return $this;
-    }
-
-    /**
-     * 获取模型对象的主键值
-     *
-     * @return mixed
-     */
-    public function getKey()
-    {
-        if ($this->key) {
-            return $this->key;
-        }
-
-        $pk = $this->getPk();
-
-        if (is_string($pk) && array_key_exists($pk, $this->data)) {
-            $this->key = $this->data[$pk];
-        } elseif (is_array($pk)) {
-            $data = [];
-            foreach ($pk as $name) {
-                if (array_key_exists($name, $this->data)) {
-                    $data[$name] = $this->data[$name];
-                }
-            }
-            $this->key = $data;
-        }
-        return $this->key ?? null;
-    }
-
-    /**
-     * 设置允许写入的字段.
-     *
-     * @param array $field 允许写入的字段
-     *
-     * @return $this
-     */
-    public function allowField(array $field)
-    {
-        $this->field = $field;
-
-        return $this;
-    }
-
-    /**
-     * 设置只读字段.
-     *
-     * @param array $field 只读字段
-     *
-     * @return $this
-     */
-    public function readonly(array $field)
-    {
-        $this->readonly = $field;
-
-        return $this;
-    }
-
-    /**
-     * 获取实际的字段名.
-     *
-     * @param string $name 字段名
+     * 获取表名（不含前后缀）.
      *
      * @return string
      */
-    public function getRealFieldName(string $name): string
+    public function getName(): string
     {
-        if ($this->convertNameToCamel || !$this->strict) {
-            return Str::snake($name);
-        }
-
-        return $name;
+        return $this->getOption('name', Str::snake(class_basename(static::class)));
     }
 
     /**
-     * 设置数据对象值
+     * 解析模型数据.
      *
-     * @param array|object $data  数据
-     * @param bool  $set   是否调用修改器
-     * @param array $allow 允许的字段名
+     * @param array|object $data 数据
      *
-     * @return $this
+     * @return array
      */
-    public function data(array | object $data, bool $set = false, array $allow = [])
+    private function parseData(array | object $data): array
     {
         if ($data instanceof self) {
             $data = $data->getData();
@@ -312,223 +126,114 @@ trait Attribute
             $data = get_object_vars($data);
         }
 
-        if (empty($data)) {
-            return $this;
-        }
-
-        // 清空数据
-        $this->data = [];
-
-        // 废弃字段
-        foreach ($this->disuse as $key) {
-            if (array_key_exists($key, $data)) {
-                unset($data[$key]);
-            }
-        }
-
-        if (!empty($allow)) {
-            $result = [];
-            foreach ($allow as $name) {
-                if (isset($data[$name])) {
-                    $result[$name] = $data[$name];
-                }
-            }
-            $data = $result;
-        }
-
-        $this->appendData($data, $set);
-
-        return $this;
-    }
-
-    /**
-     * 批量追加数据对象值
-     *
-     * @param array $data 数据
-     * @param bool  $set  是否需要进行数据处理
-     *
-     * @return $this
-     */
-    public function appendData(array $data, bool $set = false)
-    {
-        if ($set) {
-            $this->setAttrs($data);
-        } else {
-            $this->data = array_merge($this->data, $data);
-        }
-
-        return $this;
-    }
-
-    /**
-     * 刷新对象原始数据（为当前数据）.
-     *
-     * @return $this
-     */
-    public function refreshOrigin()
-    {
-        $this->origin = $this->data;
-
-        return $this;
-    }
-
-    /**
-     * 获取对象原始数据 如果不存在指定字段返回null.
-     *
-     * @param string $name 字段名 留空获取全部
-     *
-     * @return mixed
-     */
-    public function getOrigin(?string $name = null)
-    {
-        if (is_null($name)) {
-            return $this->origin;
-        }
-
-        $fieldName = $this->getRealFieldName($name);
-
-        return array_key_exists($fieldName, $this->origin) ? $this->origin[$fieldName] : null;
-    }
-
-    public function origin(array $origin)
-    {
-        $this->origin = $origin;
-        return $this;
-    }
-
-    /**
-     * 获取当前对象数据 如果不存在指定字段返回false.
-     *
-     * @param string $name 字段名 留空获取全部
-     *
-     * @throws InvalidArgumentException
-     *
-     * @return mixed
-     */
-    public function getData(?string $name = null)
-    {
-        if (is_null($name)) {
-            return $this->data;
-        }
-
-        $fieldName = $this->getRealFieldName($name);
-
-        if (array_key_exists($fieldName, $this->data)) {
-            return $this->data[$fieldName];
-        }
-
-        if (array_key_exists($fieldName, $this->relation)) {
-            return $this->relation[$fieldName];
-        }
-
-        throw new InvalidArgumentException('property not exists:' . static::class . '->' . $name);
-    }
-
-    /**
-     * 获取变化的数据 并排除只读数据.
-     *
-     * @return array
-     */
-    public function getChangedData(array $data = []): array
-    {
-        $data = $this->force ? $data : array_udiff_assoc($data, $this->origin, function ($a, $b) {
-            if ((empty($a) || empty($b)) && $a !== $b) {
-                return 1;
-            }
-
-            if ($b instanceof Raw || $b instanceof Express) {
-                return 0;
-            }
-
-            return is_object($a) || $a != $b ? 1 : 0;
-        });
-
-        // 只读字段不允许更新
-        foreach ($this->readonly as $field) {
-            if (array_key_exists($field, $data)) {
-                unset($data[$field]);
-            }
-        }
-
         return $data;
     }
 
     /**
-     * 直接设置数据对象值
+     * 设置数据字段获取器.
      *
-     * @param string $name  属性名
-     * @param mixed  $value 值
+     * @param array $attr     字段获取器定义
      *
-     * @return void
+     * @return $this
      */
-    public function set(string $name, $value): void
+    public function withFieldAttr(array $attr)
     {
-        $name = $this->getRealFieldName($name);
+        foreach ($attr as $name => $closure) {
+            $this->withAttr($name, $closure);
+        }
 
-        $this->data[$name] = $value;
-        unset($this->get[$name]);
+        return $this;
     }
 
     /**
-     * 通过修改器 批量设置数据对象值
+     * 动态设置字段获取器.
      *
-     * @param array $data 数据
+     * @param string    $name     字段名
+     * @param callable  $callback 闭包获取器
      *
-     * @return void
+     * @return $this
      */
-    public function setAttrs(array $data): void
+    public function withAttr(string $name, callable $callback)
     {
-        // 进行数据处理
-        foreach ($data as $key => $value) {
-            $this->setAttr($key, $value, $data);
-        }
+        $name = $this->getRealFieldName($name);
+
+        $this->setWeakData('with_attr', $name, $callback);
+        // 自动追加输出
+        self::$weakMap[$this]['append'][] = $name;
+        return $this;
     }
 
     /**
-     * 通过修改器 设置数据对象值
+     * 获取实际字段名.
+     * 严格模式下 完全和数据表字段对应一致（默认）
+     * 非严格模式 统一转换为snake规范（支持驼峰规范读取）
      *
-     * @param string $name  属性名
-     * @param mixed  $value 属性值
-     * @param array  $data  数据
+     * @param string $name  字段名
      *
-     * @return void
+     * @return mixed
      */
-    public function setAttr(string $name, $value, array $data = []): void
+    protected function getRealFieldName(string $name)
     {
-        if ($this->mapping) {
-            $name = array_search($name, $this->mapping) ?: $name;
+        if (false === $this->getOption('strict')) {
+            return Str::snake($name);
         }
 
-        $name = $this->getRealFieldName($name);
+        return $name;
+    }
 
-        // 检测修改器
-        $method = 'set' . Str::studly($name) . 'Attr';
-
-        if (method_exists($this, $method)) {
-            $array = $this->data;
-            $value = $this->$method($value, array_merge($this->data, $data));
-
-            if (is_null($value) && $array !== $this->data) {
-                return;
-            }
-        } elseif (!in_array($name, $this->json) && isset($this->type[$name])) {
-            // 类型转换
-            if ($this->enumStrict && is_subclass_of($this->type[$name], BackedEnum::class) && !($value instanceof BackedEnum)) {
-                throw new InvalidArgumentException('data type error: ' . $name . ' => ' . $this->type[$name]);
-            }
-            $value = $this->writeTransform($value, $this->type[$name]);
-        } elseif ($this->isRelationAttr($name)) {
-            // 关联属性
-            $this->relation[$name] = $value;
-            $this->with[$name]     = true;
-        } elseif ((array_key_exists($name, $this->origin) || empty($this->origin)) && $value instanceof Stringable) {
-            // 对象类型
-            $value = $value->__toString();
+    /**
+     * 数据读取 类型转换.
+     *
+     * @param mixed        $value 值
+     * @param string|arrau $type  要转换的类型
+     *
+     * @return mixed
+     */
+    protected function readTransform($value, string | array $type)
+    {
+        if (is_null($value)) {
+            return;
         }
 
-        // 设置数据对象属性
-        $this->data[$name] = $value;
-        unset($this->get[$name]);
+        if ($value instanceof Raw || $value instanceof Express) {
+            return $value;
+        }
+
+        if (is_array($type)) {
+            [$type, $param] = $type;
+        } elseif (str_contains($type, ':')) {
+            [$type, $param] = explode(':', $type, 2);
+        }
+
+        $typeTransform = static function (string $type, $value, $model) {
+            if (class_exists($type) && !($value instanceof $type)) {
+                if (is_subclass_of($type, Typeable::class)) {
+                    $value = $type::from($value, $model);
+                } elseif (is_subclass_of($type, FieldTypeTransform::class)) {
+                    $value = $type::get($value, $model);
+                } elseif (is_subclass_of($type, BackedEnum::class)) {
+                    $value = $type::from($value);
+                } else {
+                    // 对象类型
+                    $value = new $type($value);
+                }
+            }
+            return $value;
+        };
+
+        return match ($type) {
+            'string' => (string) $value,
+            'int'       => (int) $value,
+            'float'     => empty($param) ? (float) $value : (float) number_format($value, (int) $param, '.', ''),
+            'bool'      => (bool) $value,
+            'array'     => empty($value) ? [] : json_decode($value, true),
+            'object'    => empty($value) ? new \stdClass() : json_decode($value),
+            'json'      => $typeTransform(Json::class, $value, $this),
+            'date'      => $typeTransform(Date::class, $value, $this),
+            'datetime'  => $typeTransform(DateTime::class, $value, $this),
+            'timestamp' => $typeTransform(DateTime::class, $value, $this),
+            default     => $typeTransform($type, $value, $this),
+        };
     }
 
     /**
@@ -545,7 +250,7 @@ trait Attribute
             return;
         }
 
-        if ($value instanceof Raw) {
+        if ($value instanceof Raw || $value instanceof Express) {
             return $value;
         }
 
@@ -556,8 +261,10 @@ trait Attribute
         }
 
         $typeTransform = static function (string $type, $value, $model) {
-            if (str_contains($type, '\\') && class_exists($type)) {
-                if (is_subclass_of($type, FieldTypeTransform::class)) {
+            if (class_exists($type)) {
+                if (is_subclass_of($type, Typeable::class)) {
+                    $value = $value->value();
+                } elseif (is_subclass_of($type, FieldTypeTransform::class)) {
                     $value = $type::set($value, $model);
                 } elseif ($value instanceof BackedEnum) {
                     $value = $value->value;
@@ -565,251 +272,319 @@ trait Attribute
                     $value = $value->__toString();
                 }
             }
-
             return $value;
         };
 
         return match ($type) {
-            'string' => (string) $value,
-            'integer' => (int) $value,
-            'float' => empty($param) ? (float) $value : (float) number_format($value, (int) $param, '.', ''),
-            'boolean' => (bool) $value,
-            'timestamp' => !is_numeric($value) ? strtotime($value) : $value,
-            'datetime' => $this->formatDateTime('Y-m-d H:i:s.u', $value, true),
-            'object' => is_object($value) ? json_encode($value, JSON_FORCE_OBJECT) : $value,
-            'array' => json_encode((array) $value, !empty($param) ? (int) $param : JSON_UNESCAPED_UNICODE),
-            'json' => json_encode($value, !empty($param) ? (int) $param : JSON_UNESCAPED_UNICODE),
-            'serialize' => serialize($value),
-            default => $typeTransform($type, $value, $this),
+            'string'    => (string) $value,
+            'int'       => (int) $value,
+            'float'     => empty($param) ? (float) $value : (float) number_format($value, (int) $param, '.', ''),
+            'bool'      => (bool) $value,
+            'object'    => is_object($value) ? json_encode($value, JSON_FORCE_OBJECT) : $value,
+            'array'     => json_encode((array) $value, JSON_UNESCAPED_UNICODE),
+            'json'      => $typeTransform(Json::class, $value, $this),
+            'date'      => $typeTransform(Date::class, $value, $this),
+            'datetime'  => $typeTransform(DateTime::class, $value, $this),
+            'timestamp' => $typeTransform(DateTime::class, $value, $this),
+            default     => $typeTransform($type, $value, $this),
         };
     }
 
     /**
-     * 获取器 获取数据对象的值
+     * 刷新对象原始数据（为当前数据）.
+     *
+     * @return $this
+     */
+    public function refreshOrigin()
+    {
+        $this->setOption('origin', $this->getData());
+
+        return $this;
+    }
+
+    /**
+     * 设置主键值
+     *
+     * @param int|string $value 值
+     * @return void
+     */
+    public function setKey($value)
+    {
+        $pk = $this->getPk();
+
+        if (is_string($pk)) {
+            $this->set($pk, $value);
+        }
+    }
+
+    /**
+     * 获取主键值
+     *
+     * @return mixed
+     */
+    public function getKey()
+    {
+        $pk = $this->getPk();
+        if (is_string($pk)) {
+            return $this->get($pk);
+        }
+
+        foreach ($pk as $name) {
+            $data[$name] = $this->get($name);
+        }
+        return $data;
+    }
+
+    /**
+     * 重置模型数据.
+     *
+     * @param array $data
+     *
+     * @return void
+     */
+    public function data(array $data)
+    {
+        $this->initializeData($data);
+    }
+
+    /**
+     * 获取模型实际数据.
+     *
+     * @param string|null $name 字段名
+     * @return mixed
+     */
+    public function getData(?string $name = null)
+    {
+        if ($name) {
+            $name = $this->getRealFieldName($name);
+            return $this->getWeakData('data', $name);
+        }
+        return $this->getOption('data');
+    }
+
+    /**
+     * 设置数据对象的实际值
+     *
+     * @param string $name  名称
+     * @param mixed  $value 值
+     *
+     * @return void
+     */
+    protected function setData(string $name, $value)
+    {
+        $this->setWeakData('data', $name, $value);
+        if ($this->getWeakData('get', $name)) {
+            $this->setWeakData('get', $name, null);
+        }
+    }
+
+    /**
+     * 清空模型数据.
+     *
+     * @return $this
+     */
+    public function clear()
+    {
+        $this->setOption('data', []);
+        $this->setOption('origin', []);
+        $this->setOption('get', []);
+        $this->setOption('relation', []);
+        return $this;
+    }
+
+    /**
+     * 获取原始数据.
+     *
+     * @param string|null $name 字段名
+     * @return mixed
+     */
+    public function getOrigin(?string $name = null)
+    {
+        if ($name) {
+            $name = $this->getRealFieldName($name);
+            return $this->getWeakData('origin', $name);
+        }
+        return $this->getOption('origin');
+    }
+
+    /**
+     * 判断数据是否为空.
+     *
+     * @return bool
+     */
+    public function isEmpty(): bool
+    {
+        return empty($this->getData());
+    }
+
+    /**
+     * 设置数据对象的值 并进行类型自动转换
+     *
+     * @param string $name  名称
+     * @param mixed  $value 值
+     *
+     * @return $this
+     */
+    public function set(string $name, $value)
+    {
+        $ignore = $this->getOption('ignore', []);
+        $name   = $this->getMappingName($name);
+        $type   = $this->getFields($name);
+
+        if (in_array($name, $ignore)) {
+            // 忽略属性
+            return $this;
+        }
+
+        if (is_null($value) && is_subclass_of($type, Entity::class)) {
+            // 关联数据为空 设置一个空模型
+            $value = new $type();
+        } elseif (!($value instanceof Modelable || $value instanceof Collection)) {
+            // 类型自动转换
+            $value = $this->readTransform($value, $type);
+        }
+
+        $this->setData($name, $value);
+        return $this;
+    }
+
+    /**
+     * 使用修改器或类型自动转换处理数据（写入数据前自动调用）
+     *
+     * @param string $name  名称
+     * @param mixed  $value 值
+     * @param array  $data 所有数据
+     *
+     * @return mixed
+     */
+    private function setWithAttr(string $name, $value, array $data = [])
+    {
+        $attr   = Str::studly($name);
+        $method = 'set' . $attr . 'Attr';
+        if (method_exists($this, $attr) && $set = $this->$attr()['set'] ?? '') {
+            $value = $set($value, $data);
+        } elseif (method_exists($this, $method)) {
+            $value = $this->$method($value, $data);
+        } else {
+            // 类型转换
+            $value = $this->writeTransform($value, $this->getFields($name));
+        }
+
+        if ($value instanceof Express) {
+            // 处理运算表达式
+            $step   = $value->getStep();
+            $origin = $this->getOrigin($name);
+            $real   = match ($value->getType()) {
+                '+'         => $origin + $step,
+                '-'         => $origin - $step,
+                '*'         => $origin * $step,
+                '/'         => $origin / $step,
+                default     => $origin,
+            };
+            $this->setData($name, $real);
+        } elseif (is_scalar($value)) {
+            // 同步写入修改器或类型自动转换结果
+            $this->setData($name, $value);
+        }
+
+        return $value;
+    }
+
+    /**
+     * 获取数据对象的值（支持使用获取器）
+     *
+     * @param string $name 名称
+     * @param bool   $attr 是否使用获取器
+     *
+     * @return mixed
+     */
+    public function get(string $name, bool $attr = true)
+    {
+        $ignore = $this->getOption('ignore', []);
+        $name   = $this->getMappingName($name);
+        if (in_array($name, $ignore)) {
+            // 忽略属性
+            return null;
+        }
+
+        if ($attr && $value = $this->getWeakData('get', $name)) {
+            // 已经输出的数据直接返回
+            return $value;
+        }
+
+        if (!array_key_exists($name, $this->getOption('data'))) {
+            // 动态获取关联数据
+            $value = $this->getRelationData($name) ?: null;
+        } else {
+            $value = $this->getData($name);
+        }
+
+        if ($attr) {
+            // 通过获取器输出
+            $value = $this->getWithAttr($name, $value, $this->getData());
+            $this->setWeakData('get', $name, $value);
+        }
+
+        return $value;
+    }
+
+    /**
+     * 获取映射字段
      *
      * @param string $name 名称
      *
-     * @throws InvalidArgumentException
-     *
-     * @return mixed
+     * @return string
      */
-    public function getAttr(string $name)
+    protected function getMappingName(string $name): string
     {
-        try {
-            $relation = false;
-            if ($this->mapping) {
-                $name = array_search($name, $this->mapping) ?: $name;
-            }
-            $value = $this->getData($name);
-        } catch (InvalidArgumentException $e) {
-            $relation = $this->isRelationAttr($name);
-            $value    = null;
+        $mapping = $this->getOption('mapping');
+        if (!empty($mapping)) {
+            $name = array_search($name, $mapping) ?: $name;
         }
-
-        return $this->getValue($name, $value, $relation);
+        return $this->getRealFieldName($name);
     }
 
     /**
-     * 获取经过获取器处理后的数据对象的值
+     * 处理数据对象的值（经过获取器和类型转换）
      *
-     * @param string      $name     字段名称
-     * @param mixed       $value    字段值
-     * @param bool|string $relation 是否为关联属性或者关联名
-     *
-     * @throws InvalidArgumentException
+     * @param string $name 名称
+     * @param mixed  $value 值
+     * @param array  $data 所有数据
      *
      * @return mixed
      */
-    protected function getValue(string $name, $value, bool | string $relation = false)
+    private function getWithAttr(string $name, $value, array $data = [])
     {
-        // 检测属性获取器
-        $fieldName = $this->getRealFieldName($name);
-
-        if (array_key_exists($fieldName, $this->get)) {
-            return $this->get[$fieldName];
-        }
-
-        $method = 'get' . Str::studly($name) . 'Attr';
-        if (isset($this->withAttr[$fieldName])) {
-            if ($relation) {
-                $value = $this->getRelationValue($relation);
-            }
-
-            if (in_array($fieldName, $this->json) && is_array($this->withAttr[$fieldName])) {
-                $value = $this->getJsonValue($fieldName, $value);
-            } else {
-                $closure = $this->withAttr[$fieldName];
-                if ($closure instanceof Closure) {
-                    $value = $closure($value, $this->data, $this);
-                }
-            }
+        $attr     = Str::studly($name);
+        $method   = 'get' . $attr . 'Attr';
+        $withAttr = $this->getWeakData('with_attr', $name);
+        if ($withAttr) {
+            // 动态获取器
+            $value = $withAttr($value, $data, $this);
+        } elseif (method_exists($this, $attr) && $get = $this->$attr()['get'] ?? '') {
+            // 属性器
+            $value = $get($value, $data);
         } elseif (method_exists($this, $method)) {
-            if ($relation) {
-                $value = $this->getRelationValue($relation);
-            }
-
-            $value = $this->$method($value, $this->data);
-        } elseif (!in_array($fieldName, $this->json) && isset($this->type[$fieldName])) {
-            // 类型转换
-            $value = $this->readTransform($value, $this->type[$fieldName]);
-        } elseif ($this->autoWriteTimestamp && in_array($fieldName, [$this->createTime, $this->updateTime])) {
-            $value = $this->getTimestampValue($value);
-        } elseif ($relation) {
-            $value = $this->getRelationValue($relation);
-            // 保存关联对象值
-            $this->relation[$name] = $value;
+            // 获取器
+            $value = $this->$method($value, $data);
+        } elseif ($value instanceof Typeable || is_subclass_of($value, EnumTransform::class)) {
+            // 类型自动转换
+            $value = $value->value();
         }
-
-        $this->get[$fieldName] = $value;
-
         return $value;
     }
 
     /**
-     * 获取JSON字段属性值
+     * 设置忽略属性.
      *
-     * @param string $name  属性名
-     * @param mixed  $value JSON数据
-     *
-     * @return mixed
-     */
-    protected function getJsonValue(string $name, $value)
-    {
-        if (is_null($value)) {
-            return $value;
-        }
-
-        foreach ($this->withAttr[$name] as $key => $closure) {
-            if ($this->jsonAssoc) {
-                $value[$key] = $closure($value[$key] ?? '', $value);
-            } else {
-                $value->$key = $closure($value->$key ?? '', $value);
-            }
-        }
-
-        return $value;
-    }
-
-    /**
-     * 获取关联属性值
-     *
-     * @param string $relation 关联名
-     *
-     * @return mixed
-     */
-    protected function getRelationValue(string $relation)
-    {
-        $modelRelation = $this->$relation();
-
-        return $modelRelation instanceof Relation ? $this->getRelationData($modelRelation) : null;
-    }
-
-    /**
-     * 数据读取 类型转换.
-     *
-     * @param mixed        $value 值
-     * @param string|array $type  要转换的类型
-     *
-     * @return mixed
-     */
-    protected function readTransform($value, string | array $type)
-    {
-        if (is_null($value)) {
-            return;
-        }
-
-        if (is_array($type)) {
-            [$type, $param] = $type;
-        } elseif (str_contains($type, ':')) {
-            [$type, $param] = explode(':', $type, 2);
-        }
-
-        $call = function ($value) {
-            try {
-                $value = unserialize($value);
-            } catch (\Exception $e) {
-                $value = null;
-            }
-            return $value;
-        };
-
-        $typeTransform = static function (string $type, $value, $model) {
-            if (str_contains($type, '\\') && class_exists($type)) {
-                if (is_subclass_of($type, FieldTypeTransform::class)) {
-                    $value = $type::get($value, $model);
-                } elseif (is_subclass_of($type, BackedEnum::class)) {
-                    $value = $type::from($value);
-                    if (is_subclass_of($type, EnumTransform::class)) {
-                        $value = $value->value();
-                    } elseif ($model->enumReadName) {
-                        $method = $model->enumReadName;
-                        $value  = is_string($method) ? $value->$method() : $value->name;
-                    }
-                } else {
-                    // 对象类型
-                    $value = new $type($value);
-                }
-            }
-
-            return $value;
-        };
-
-        return match ($type) {
-            'string' => (string) $value,
-            'integer' => (int) $value,
-            'float' => empty($param) ? (float) $value : (float) number_format($value, (int) $param, '.', ''),
-            'boolean' => (bool) $value,
-            'timestamp' => !is_null($value) ? $this->formatDateTime(!empty($param) ? $param : $this->dateFormat, $value, true) : null,
-            'datetime' => !is_null($value) ? $this->formatDateTime(!empty($param) ? $param : $this->dateFormat, $value) : null,
-            'json' => json_decode($value, true),
-            'array' => empty($value) ? [] : json_decode($value, true),
-            'object' => empty($value) ? new \stdClass() : json_decode($value),
-            'serialize' => $call($value),
-            default => $typeTransform($type, $value, $this),
-        };
-    }
-
-    /**
-     * 设置数据字段获取器.
-     *
-     * @param string|array $name     字段名
-     * @param Closure     $callback 闭包获取器
+     * @param array $ignore 忽略属性列表
      *
      * @return $this
      */
-    public function withFieldAttr(string | array $name, ?Closure $callback = null)
+    public function ignore(array $ignore)
     {
-        if (is_array($name)) {
-            foreach ($name as $key => $val) {
-                $this->withFieldAttr($key, $val);
-            }
-        } else {
-            $name = $this->getRealFieldName($name);
-            $this->append([$name], true);
-
-            if (str_contains($name, '.')) {
-                [$name, $key] = explode('.', $name);
-
-                $this->withAttr[$name][$key] = $callback;
-            } else {
-                $this->withAttr[$name] = $callback;
-            }
-        }
+        $this->setOption('ignore', $ignore);
 
         return $this;
-    }
-
-    /**
-     * 设置枚举类型自动读取数据方式
-     * true 表示使用name值返回
-     * 字符串 表示使用枚举类的方法返回
-     *
-     * @return $this
-     */
-    public function withEnumRead(bool | string $method = true)
-    {
-        $this->enumReadName = $method;
-
-        return $this;
-    }
+    }    
 }
