+++ src/model/concern/OptimLock.php
@@ -22,64 +22,66 @@ trait OptimLock
 {
     protected function getOptimLockField()
     {
-        return property_exists($this, 'optimLock') && isset($this->optimLock) ? $this->optimLock : 'lock_version';
+        return $this->getOption('optim_lock_field') ?? 'lock_version';
     }
 
     /**
      * 数据检查.
-     *
+     * @param array $data 数据
+     * @param bool  $isUpdate 是否更新
      * @return void
      */
-    protected function checkData(): void
+    protected function checkData(array &$data, bool $isUpdate): void
     {
-        $this->isExists() ? $this->updateLockVersion() : $this->recordLockVersion();
+        $isUpdate ? $this->updateLockVersion($data) : $this->recordLockVersion($data);
     }
 
     /**
      * 记录乐观锁
      *
+     * @param array $data 数据
      * @return void
      */
-    protected function recordLockVersion(): void
+    protected function recordLockVersion(array &$data): void
     {
         $optimLock = $this->getOptimLockField();
 
-        if ($optimLock) {
-            $this->set($optimLock, 0);
-        }
+        $this->setData($optimLock, 0);
+        $data[$optimLock] = 0;
     }
 
     /**
      * 更新乐观锁
      *
+     * @param array $data 数据
      * @return void
      */
-    protected function updateLockVersion(): void
+    protected function updateLockVersion(array &$data): void
     {
         $optimLock = $this->getOptimLockField();
+        $lockVer   = $this->getOrigin($optimLock);
 
-        if ($optimLock && $lockVer = $this->getOrigin($optimLock)) {
-            // 更新乐观锁
-            $this->set($optimLock, $lockVer + 1);
-        }
+        $this->setData($optimLock, $lockVer + 1);
+        $data[$optimLock] = $lockVer + 1;
     }
 
-    public function getWhere()
+    public function getDbWhere($where)
     {
-        $where      = parent::getWhere();
+        $db = $this->db();
+        // 检查条件
+        if (!empty($where)) {
+            $db->where($where);
+        }
         $optimLock  = $this->getOptimLockField();
-
-        if ($optimLock && $lockVer = $this->getOrigin($optimLock)) {
-            $where[] = [$optimLock, '=', $lockVer];
+        $lockVer    = $this->getOrigin($optimLock);
+        $pk         = $this->getPk();
+        if (is_array($pk)) {
+            $db->where($this->getKey());
+        } else {
+            $db->where($pk, '=', $this->getKey());
         }
+        $db->where($optimLock, '=', $lockVer);
 
-        return $where;
-    }
-
-    protected function checkResult($result): void
-    {
-        if (!$result) {
-            throw new Exception('record has update');
-        }
+        return $db;
     }
 }
